refactor(backend): stop Any leaking into declared return types
`unsound-return-statement` catches a return whose value is assignable to the
declared type but not a subtype of it — in practice, `Any` from an untyped
boundary being waved through as if it were checked. 13 sites, down from 24
before the previous commit made the adaptix loader generic.

Most are now genuinely narrowed rather than suppressed. `get_constraint_name`
reads three attributes off a driver object with no stubs and returned whichever
one was truthy as `str | None`; isinstance guards are what make that signature
true. VK's message id is `int()`-checked, the mailing counters likewise, the
OAuth url stringified, and an empty-headers literal moved to a typed constant
(a bare `{}` infers `dict[str, Unknown]`).

Four stay suppressed with a reason: SQLAlchemy types mapped_column() and
column_property() as Any-ish builders, and asyncpg ships no type information
at all.

Separately, `just backend-check` no longer passes `--unsafe-fixes`. Ruff marks
a fix unsafe when it can change behaviour or drop a comment, so a gate carrying
it rewrites the code of whoever ran it to check their work, which is the
opposite of what a gate is for. The flag stays available by hand for a
deliberate bulk pass. Pre-commit already used safe fixes only and CI was
already check-only.

+++ backend/src/fanfan/adapters/api/base.py
@@ -79,4 +79,4 @@ def _is_retryable(error: httpx2.TransportError | httpx2.HTTPStatusError) -> bool
 
 def _backoff_delay(attempt: int) -> float:
     capped = min(RETRY_MAX_DELAY, RETRY_BASE_DELAY * 2 ** (attempt - 1))
-    return capped * random.uniform(0.5, 1.5)  # noqa: S311 — jitter, not a security draw
+    return float(capped * random.uniform(0.5, 1.5))  # noqa: S311 — jitter, not a draw
